paginacao backend

+++ ProjetoDotinha2/Models/MatchModel.cs
@@ -4,7 +4,7 @@
     {
         public List<PlayerModel> Players { get; set; }
 
-        public int matchId { get; set; }
+        public long matchId { get; set; }
 
         public int patch { get; set; }
     }
